Moved AclsClient class to more appropriate package
