+++ aclslib/src/main/java/au/edu/uq/cmm/acslib/authenticator/Authenticator.java
@@ -2,6 +2,7 @@ package au.edu.uq.cmm.acslib.authenticator;
 
 import org.apache.log4j.Logger;
 
+import au.edu.uq.cmm.aclslib.message.AclsClient;
 import au.edu.uq.cmm.aclslib.message.AclsProtocolException;
 import au.edu.uq.cmm.aclslib.message.LoginRequest;
 import au.edu.uq.cmm.aclslib.message.Request;
@@ -9,7 +10,6 @@ import au.edu.uq.cmm.aclslib.message.RequestType;
 import au.edu.uq.cmm.aclslib.message.Response;
 import au.edu.uq.cmm.aclslib.message.SimpleRequest;
 import au.edu.uq.cmm.aclslib.message.YesNoResponse;
-import au.edu.uq.cmm.aclslib.proxy.AclsClient;
 import au.edu.uq.cmm.aclslib.server.Configuration;
 
 public class Authenticator {
